+++ obtain_new_ids.py
@@ -27,6 +27,7 @@ class IDFinder():
         self.thoth = ThothClient()
         self.thoth_ids = []
         self.work_statuses = None
+        self.work_types = None
         self.publishers = None
         self.order = None
         self.updated_at_with_relations = None
@@ -82,6 +83,8 @@ class IDFinder():
         """
         # Default: all active (published) works listed in Thoth (from the selected publishers).
         self.work_statuses = '[ACTIVE]'
+        # Default: all work types included except for chapters (from the selected publishers).
+        self.work_types = '[MONOGRAPH, EDITED_BOOK, JOURNAL_ISSUE, TEXTBOOK, BOOK_SET]'
         # Start with the most recent, so that we can disregard everything else
         # as soon as we hit the first work published earlier than the desired date range.
         self.order = '{field: PUBLICATION_DATE, direction: DESC}'
@@ -112,10 +115,11 @@ class IDFinder():
         # we can simply construct a standard query filtering by publication date
         offset = 0
         while True:
-            next_batch = self.thoth.books(
+            next_batch = self.thoth.works(
                 limit=1,
                 offset=offset,
                 work_statuses=self.work_statuses,
+                work_types=self.work_types,
                 order=self.order,
                 publishers=self.publishers,
                 updated_at_with_relations=self.updated_at_with_relations,
@@ -183,7 +187,7 @@ class CrossrefIDFinder(IDFinder):
         DELAY_BUFFER_HRS = 0.25
 
         # Target: all works listed in Thoth (from the selected publishers) which are
-        # Active, and which have been updated since the last deposit.
+        # Active or Forthcoming, and which have been updated since the last deposit.
         # Use UTC, as GitHub Actions scheduling runs in UTC.
         current_time = datetime.now(UTC)
         last_deposit_time = current_time - \
@@ -191,11 +195,22 @@ class CrossrefIDFinder(IDFinder):
         last_deposit_time_str = datetime.strftime(
             last_deposit_time, "%Y-%m-%dT%H:%M:%SZ")
 
-        self.work_statuses = '[ACTIVE]'
+        self.work_statuses = '[ACTIVE, FORTHCOMING]'
         # Start with the most recently updated
         self.order = '{field: UPDATED_AT_WITH_RELATIONS, direction: DESC}'
         self.updated_at_with_relations = '{{timestamp: "{}", expression: GREATER_THAN}}'.format(
             last_deposit_time_str)
+
+    def post_process(self):
+        """
+        Exclude from the results any Forthcoming works which don't yet have
+        both a DOI and publication date.
+        """
+        for id in reversed(self.thoth_ids):
+            work = self.thoth.work_by_id(id)
+            if work.workStatus == 'FORTHCOMING':
+                if not work.doi or not work.publicationDate:
+                    self.thoth_ids.remove(id)
 
 
 class InternetArchiveIDFinder(IDFinder):
@@ -292,6 +307,36 @@ class OapenIDFinder(IDFinder):
         self.get_thoth_ids_iteratively(previous_week_start, previous_week_end)
 
 
+class BKCIIDFinder(IDFinder):
+    """Logic for retrieving work IDs which is specific to Clarivate Web of Science Book Citation Index (BKCI) dissemination"""
+
+    def get_query_parameters(self):
+        """Construct Thoth work ID query parameters depending on Clarivate BKCI-specific requirements"""
+        # Target: all active (published) works listed in Thoth (from the selected publishers), except for textbooks
+        self.work_statuses = '[ACTIVE]'
+        self.work_types = '[MONOGRAPH, EDITED_BOOK, JOURNAL_ISSUE, BOOK_SET]'
+        # Start with the most recent, so that we can disregard everything else
+        # as soon as we hit the first work published earlier than the desired date range.
+        self.order = '{field: PUBLICATION_DATE, direction: DESC}'
+        self.updated_at_with_relations = None
+
+    def get_thoth_ids(self):
+        """Query Thoth GraphQL API with relevant parameters to retrieve required work IDs"""
+        # TODO Once https://github.com/thoth-pub/thoth/issues/486 is completed,
+        # we can remove this overriding method and simply construct a standard query
+        # filtering by publication date
+
+        # In addition to the conditions of the query parameters, we need to filter the results
+        # to obtain only works with a publication date within the previous calendar month.
+        # The schedule for finding and depositing newly published works is once monthly
+        # (a few days after the start of the month, to allow for delays in updating records).
+        current_date = datetime.now(UTC).date()
+        current_month_start = current_date.replace(day=1)
+        previous_month_end = current_month_start - timedelta(days=1)
+        previous_month_start = previous_month_end.replace(day=1)
+
+        self.get_thoth_ids_iteratively(previous_month_start, previous_month_end)
+
 class OapenLocationsIDFinder(IDFinder):
     """
     Helper class for workflow which updates Thoth records with newly-registered
@@ -372,10 +417,12 @@ if __name__ == '__main__':
                 id_finder = OapenIDFinder()
             case 'Figshare' | 'Zenodo' | 'CUL':
                 id_finder = CatchupIDFinder()
+            case 'BKCI':
+                id_finder = BKCIIDFinder()
             case _:
                 logging.error(
                     'Platform must be one of InternetArchive, Crossref, Figshare, '
-                    'Zenodo, CUL, GooglePlay or OAPEN')
+                    'Zenodo, CUL, GooglePlay, BKCI or OAPEN')
                 sys.exit(1)
 
     id_finder.run()